5/30 updates

+++ Mask_RCNN/mrcnn/visualize.py
@@ -163,7 +163,7 @@ def display_instances(image, boxes, masks, class_ids, class_names,
             verts = np.fliplr(verts) - 1
             p = Polygon(verts, facecolor="none", edgecolor=color)
             ax.add_patch(p)
-    ax.imshow(masked_image.astype(np.uint8))
+    ax.imshow(masked_image.astype(np.uint32))
     if auto_show:
         plt.show()
